requirements.txt
streamlit
openai
tqdm
python-dotenv
streamlit-authenticator
pandas
llama-index
llama-index-vector-stores-pinecone
llama-index-llms-openai
protobuf==4.25.5
supabase
pinecone[grpc]